+++ src/drawing/area.rs
@@ -121,10 +121,9 @@ impl<DB: DrawingBackend> From<DB> for DrawingArea<DB, Shift> {
     fn from(backend: DB) -> Self {
         Self::with_rc_cell(Rc::new(RefCell::new(backend)))
     }
-
 }
 
-impl <'a, DB: DrawingBackend> From<&'a Rc<RefCell<DB>>> for DrawingArea<DB, Shift> {
+impl<'a, DB: DrawingBackend> From<&'a Rc<RefCell<DB>>> for DrawingArea<DB, Shift> {
     fn from(backend: &'a Rc<RefCell<DB>>) -> Self {
         Self::with_rc_cell(backend.clone())
     }
@@ -171,7 +170,6 @@ impl<DB: DrawingBackend, X: Ranged, Y: Ranged> DrawingArea<DB, RangedCoord<X, Y>
 }
 
 impl<DB: DrawingBackend, CT: CoordTranslate> DrawingArea<DB, CT> {
-
     /// Get the left upper conner of this area in the drawing backend
     pub fn get_base_pixel(&self) -> BackendCoord {
         (self.rect.x0, self.rect.y0)
@@ -275,7 +273,6 @@ impl<DB: DrawingBackend, CT: CoordTranslate> DrawingArea<DB, CT> {
 }
 
 impl<DB: DrawingBackend> DrawingArea<DB, Shift> {
-
     fn with_rc_cell(backend: Rc<RefCell<DB>>) -> Self {
         let (x1, y1) = RefCell::borrow(backend.borrow()).get_size();
         Self {